django_api/movies/app/requirements.txt
django==4.2.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-split-settings==1.2.0
django_debug_toolbar==4.2.0
uWSGI==2.0.23
flake8==6.1.0
pre-commit==3.5.0
flake8-bandit==4.1.1
bandit==1.7.5
django-cors-headers==4.3.1
django_extensions==3.2.3.
